+++ CHANGELOG.md
@@ -2254,3 +2254,10 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 * **node**
 
   1.  Added [fiddles/node/fiddle-0040-DropBoxSdk](fiddles/node/fiddle-0040-DropBoxSdk)
+
+
+### 333_node_41 ~ [Issue 333](https://github.com/bradyhouse/house/issues/333)
+
+* **node**
+
+    1.  Added [fiddles/node/fiddle-0041-SvgWriter](fiddles/node/fiddle-0041-SvgWriter)